refactor: simplify nfs client and nfs server (#771)
