+++ src/nfs/nfs_client_impl.cpp
@@ -135,15 +135,12 @@ void nfs_client_impl::begin_remote_copy(std::shared_ptr<remote_copy_request> &rc
     req->nfs_task = nfs_task;
     req->is_finished = false;
 
-    get_file_size(req->file_size_req,
-                  [=](error_code err, get_file_size_response &&resp) {
-                      end_get_file_size(err, std::move(resp), req);
-                  },
-                  std::chrono::milliseconds(FLAGS_rpc_timeout_ms),
-                  0,
-                  0,
-                  0,
-                  req->file_size_req.source);
+    async_nfs_get_file_size(req->file_size_req,
+                            [=](error_code err, get_file_size_response &&resp) {
+                                end_get_file_size(err, std::move(resp), req);
+                            },
+                            std::chrono::milliseconds(FLAGS_rpc_timeout_ms),
+                            req->file_size_req.source);
 }
 
 void nfs_client_impl::end_get_file_size(::dsn::error_code err,
@@ -284,22 +281,20 @@ void nfs_client_impl::continue_copy()
                 copy_req.source_dir = ureq->file_size_req.source_dir;
                 copy_req.overwrite = ureq->file_size_req.overwrite;
                 copy_req.is_last = req->is_last;
-                req->remote_copy_task = copy(copy_req,
-                                             [=](error_code err, copy_response &&resp) {
-                                                 end_copy(err, std::move(resp), req);
-                                                 // reset task to release memory quickly.
-                                                 // should do this after end_copy() done.
-                                                 if (req->is_ready_for_write) {
-                                                     ::dsn::task_ptr tsk;
-                                                     zauto_lock l(req->lock);
-                                                     tsk = std::move(req->remote_copy_task);
-                                                 }
-                                             },
-                                             std::chrono::milliseconds(FLAGS_rpc_timeout_ms),
-                                             0,
-                                             0,
-                                             0,
-                                             req->file_ctx->user_req->file_size_req.source);
+                req->remote_copy_task =
+                    async_nfs_copy(copy_req,
+                                   [=](error_code err, copy_response &&resp) {
+                                       end_copy(err, std::move(resp), req);
+                                       // reset task to release memory quickly.
+                                       // should do this after end_copy() done.
+                                       if (req->is_ready_for_write) {
+                                           ::dsn::task_ptr tsk;
+                                           zauto_lock l(req->lock);
+                                           tsk = std::move(req->remote_copy_task);
+                                       }
+                                   },
+                                   std::chrono::milliseconds(FLAGS_rpc_timeout_ms),
+                                   req->file_ctx->user_req->file_size_req.source);
             } else {
                 --ureq->concurrent_copy_count;
                 --_concurrent_copy_request_count;
